+++ modules/docker.nix
@@ -1,15 +1,11 @@
 {
   nixosModule = { user, pkgs, ... }: {
     environment.systemPackages = builtins.attrValues {
-      inherit (pkgs) docker-client;
+      # Uses podman-compose instead of docker-compose
+      inherit (pkgs) arion;
     };
 
     virtualisation.podman.enable = true;
-
-    # Necessary for `arion`
-    virtualisation.podman.dockerSocket.enable = true;
-    virtualisation.podman.defaultNetwork.dnsname.enable = true;
-
-    users.users.${user}.extraGroups = [ "podman" ];
+    virtualisation.podman.dockerCompat = true;
   };
 }